Web Scraping in Rvest
Data and information on the web is growing exponentially. All of us today use Google as our first source of knowledge – be it about finding reviews about a place to understanding a new term. All this information is available on the web already.

With the amount of data available over the web, it opens new horizons of possibility for a Data Scientist. I strongly believe web scraping is a must have skill for any data scientist. In today’s world, all the data that you need is already available on the internet – the only thing limiting you from using it is the ability to access it. With the help of this article, you will be able to overcome that barrier as well.

Most of the data available over the web is not readily available. It is present in an unstructured format (HTML format) and is not downloadable. Therefore, it requires knowledge & expertise to use this data to eventually build a useful model.